+++ pages/api/search.ts
@@ -1,5 +1,4 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import rankEval from "../../rank_eval.json";
 import searchTemplates from "../../search_templates.json";
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
